Separate DuplexMatrixScanner into its own file.
